+++ src/commands/validate-tasks.ts
@@ -1,14 +1,15 @@
 import { Command } from 'commander';
-import { BaseCliCommand, ICommandArgs } from './base-command';
+import { BaseCliCommand } from './base-command';
 import { BuildConfiguration } from '~build-tasks/build-configuration';
 import { BuildRunner } from '~build-tasks/build-runner';
+import { IPerformTasksCommandArgs } from './perform-tasks';
 
 const commandName = 'validate-tasks <templateFile>';
 const commandDescription = 'Will validate the tasks file, including configured tasks';
 
-export class ValidateTasksCommand extends BaseCliCommand<IValidateTasksCommandArgs> {
+export class ValidateTasksCommand extends BaseCliCommand<IPerformTasksCommandArgs> {
 
-    public static async Perform(command: IValidateTasksCommandArgs): Promise<void> {
+    public static async Perform(command: IPerformTasksCommandArgs): Promise<void> {
         const x = new ValidateTasksCommand();
         await x.performCommand(command);
     }
@@ -18,17 +19,15 @@ export class ValidateTasksCommand extends BaseCliCommand<IValidateTasksCommandAr
     }
 
     public addOptions(command: Command): void {
+        command.option('--organization-file [organization-file]', 'organization file used for organization bindings');
+
         super.addOptions(command);
     }
 
-    public async performCommand(command: IValidateTasksCommandArgs): Promise<void> {
+    public async performCommand(command: IPerformTasksCommandArgs): Promise<void> {
         const tasksFile = command.tasksFile;
         const config = new BuildConfiguration(tasksFile);
         const validationTasks = config.enumValidationTasks(command);
         await BuildRunner.RunValidationTasks(validationTasks, 1, 999);
     }
-}
-
-export interface IValidateTasksCommandArgs extends ICommandArgs {
-    tasksFile: string;
 }